# openapi.yaml
openapi: 3.0.3
info:
  title: org-cost-api
  description: |
    Read-only AWS organization cost explorer API.

    Auth-exempt routes: `/api/health`, `/api/ready`, `/api/meta`.
    When `api_token_env` is configured, all other routes require `Authorization: Bearer <token>`.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://org-cost-api-demo.fly.dev
    description: Public synthetic demo (fixture data only)
security:
  - bearerAuth: []
paths:
  /api/health:
    get:
      summary: Liveness probe
      description: Auth-exempt. Use for Kubernetes liveness / LB process checks.
      security: []
      responses:
        "200":
          description: Process is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /api/ready:
    get:
      summary: Readiness probe
      description: >
        Auth-exempt. Returns ready=true when account clients are loaded.
        billing_check is "ok" after a payer STS/profile probe, "failed" when billing
        is configured but the probe fails (ready=false), or "skipped" when no billing
        is configured.
      security: []
      responses:
        "200":
          description: Ready to serve traffic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyStatus"
        "503":
          description: Not ready (billing STS check failed or no accounts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyStatus"
  /api/meta:
    get:
      summary: Deployment metadata
      description: Auth-exempt. Used by the UI demo banner and MCP health checks.
      security: []
      responses:
        "200":
          description: Meta payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaResponse"
  /api/accounts:
    get:
      summary: List configured accounts
      responses:
        "200":
          description: Account id/name list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/dashboard:
    get:
      summary: Organization dashboard
      parameters:
        - $ref: "#/components/parameters/refresh"
      responses:
        "200":
          description: Dashboard payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/trends:
    get:
      summary: Period-over-period trends
      parameters:
        - $ref: "#/components/parameters/refresh"
        - $ref: "#/components/parameters/period"
      responses:
        "200":
          description: Trends payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/suggestions:
    get:
      summary: Cost optimization suggestions
      parameters:
        - $ref: "#/components/parameters/refresh"
        - $ref: "#/components/parameters/period"
      responses:
        "200":
          description: Suggestions payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/report:
    get:
      summary: Dashboard + trends + suggestions bundle
      parameters:
        - $ref: "#/components/parameters/refresh"
        - $ref: "#/components/parameters/period"
      responses:
        "200":
          description: Report bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/account-costs:
    get:
      summary: Per-account costs
      parameters:
        - name: account
          in: query
          required: true
          description: Account name or 12-digit account id
          schema:
            type: string
        - $ref: "#/components/parameters/refresh"
      responses:
        "200":
          description: Account costs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountCostsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/ask:
    post:
      summary: Natural-language cost question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
      responses:
        "200":
          description: Ask response with answer markdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/service-detail:
    get:
      summary: Service drill-down for one account
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
        - name: service
          in: query
          required: true
          description: Cost Explorer SERVICE name (exact match)
          schema:
            type: string
        - name: start
          in: query
          schema:
            type: string
            format: date
        - name: end
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Service detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/service-tag-delta:
    get:
      summary: Tag-bucket spend change explanation
      parameters:
        - name: service
          in: query
          required: true
          schema:
            type: string
        - name: account_id
          in: query
          schema:
            type: string
        - name: start
          in: query
          schema:
            type: string
            format: date
        - name: end
          in: query
          schema:
            type: string
            format: date
        - name: tag_key
          in: query
          schema:
            type: string
            default: Name
        - name: top_accounts
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Tag delta response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceTagDeltaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/service-tag-totals:
    get:
      summary: Top tag buckets by spend for a service
      parameters:
        - name: service
          in: query
          required: true
          schema:
            type: string
        - name: account_id
          in: query
          schema:
            type: string
        - name: start
          in: query
          schema:
            type: string
            format: date
        - name: end
          in: query
          schema:
            type: string
            format: date
        - name: tag_key
          in: query
          schema:
            type: string
            default: Name
        - name: top_buckets
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Tag totals response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceTagTotalsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    refresh:
      name: refresh
      in: query
      description: >
        Force live AWS Cost Explorer fetch. Truthy when set to "1", "true", or any
        value except "0", "false", or empty (uses cache when available).
      schema:
        type: string
    period:
      name: period
      in: query
      description: >
        Cost window preset. "30d" (default) uses cost_lookback_days trailing window;
        "mtd" is calendar month-to-date. Prior period is an equal apples-to-apples
        comparison (immediate prior window for 30d; same days last month for mtd).
      schema:
        type: string
        enum: [30d, mtd, lookback]
        default: 30d
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request (missing required query param or invalid JSON)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Unknown account or resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    MetaResponse:
      type: object
      required: [demo]
      properties:
        demo:
          type: boolean
          description: true when serving synthetic fixture data
    ReadyStatus:
      type: object
      properties:
        ready:
          type: boolean
        accounts_configured:
          type: integer
        snapshot_count:
          type: integer
        billing_check:
          type: string
          enum: [ok, skipped, failed]
        billing_error:
          type: string
        accounts_check:
          type: string
          enum: [ok, skipped, failed]
        accounts_error:
          type: string
    AccountEntry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    PeriodSummary:
      type: object
      properties:
        start:
          type: string
          format: date
        end:
          type: string
          format: date
        days:
          type: integer
    OrgServiceDriver:
      type: object
      properties:
        service:
          type: string
        amount:
          type: number
        unit:
          type: string
        percent:
          type: number
        top_accounts:
          type: array
          items:
            $ref: "#/components/schemas/OrgServiceAccountShare"
    OrgServiceAccountShare:
      type: object
      properties:
        account_id:
          type: string
        account_name:
          type: string
        amount:
          type: number
        percent:
          type: number
    ConsolidatedTotals:
      type: object
      properties:
        org_total:
          type: number
        ec2_other_cost:
          type: number
        unit:
          type: string
        volume_count:
          type: integer
        volume_available_count:
          type: integer
        volume_available_gib:
          type: number
        volume_size_gib:
          type: number
        snapshot_count:
          type: integer
        snapshot_size_gib:
          type: number
        account_count:
          type: integer
    AccountDashboard:
      type: object
      properties:
        account_id:
          type: string
        account_name:
          type: string
        error:
          type: string
        costs:
          type: object
          additionalProperties: true
        volumes:
          type: object
          additionalProperties: true
        snapshots:
          type: object
          additionalProperties: true
    DashboardResponse:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        start:
          type: string
          format: date
        end:
          type: string
          format: date
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountDashboard"
        totals:
          $ref: "#/components/schemas/ConsolidatedTotals"
        top_services:
          type: array
          items:
            $ref: "#/components/schemas/OrgServiceDriver"
        cur_enabled:
          type: boolean
        cur_note:
          type: string
    ServiceTrend:
      type: object
      properties:
        service:
          type: string
        display_name:
          type: string
        current_usd:
          type: number
        prior_usd:
          type: number
        change_usd:
          type: number
        change_percent:
          type: number
        direction:
          type: string
          enum: [up, down, new, stable]
        current_share_pct:
          type: number
    AccountTrend:
      type: object
      properties:
        account_id:
          type: string
        account_name:
          type: string
        current_usd:
          type: number
        prior_usd:
          type: number
        change_usd:
          type: number
        change_percent:
          type: number
        direction:
          type: string
          enum: [up, down, new, stable]
        current_share_pct:
          type: number
    ChangeSummary:
      type: object
      properties:
        current_usd:
          type: number
        prior_usd:
          type: number
        change_usd:
          type: number
        change_percent:
          type: number
    TrendsResponse:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        period:
          type: string
          description: Active period preset (30d or mtd)
          enum: [30d, mtd]
        current_period:
          $ref: "#/components/schemas/PeriodSummary"
        prior_period:
          $ref: "#/components/schemas/PeriodSummary"
        prior_source:
          type: string
        ce_calls_used:
          type: integer
        org_total:
          $ref: "#/components/schemas/ChangeSummary"
        service_trends:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTrend"
        top_increases:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTrend"
        top_decreases:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTrend"
        account_trends:
          type: array
          items:
            $ref: "#/components/schemas/AccountTrend"
        top_account_increases:
          type: array
          items:
            $ref: "#/components/schemas/AccountTrend"
        top_account_decreases:
          type: array
          items:
            $ref: "#/components/schemas/AccountTrend"
        history_note:
          type: string
        snapshot_count:
          type: integer
        refresh_allowed:
          type: boolean
    Suggestion:
      type: object
      properties:
        id:
          type: string
        priority:
          type: integer
        category:
          type: string
        title:
          type: string
        detail:
          type: string
        estimated_monthly_usd:
          type: number
        actions:
          type: array
          items:
            type: string
        account:
          type: string
        service:
          type: string
    SuggestionsResponse:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        period:
          type: string
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/Suggestion"
        summary:
          type: string
        ce_calls_used:
          type: integer
        data_sources:
          type: array
          items:
            type: string
    ReportResponse:
      type: object
      properties:
        dashboard:
          $ref: "#/components/schemas/DashboardResponse"
        trends:
          $ref: "#/components/schemas/TrendsResponse"
        suggestions:
          $ref: "#/components/schemas/SuggestionsResponse"
        ce_calls_used:
          type: integer
        refresh_allowed:
          type: boolean
    AccountCostsResponse:
      type: object
      properties:
        account_id:
          type: string
        account_name:
          type: string
        period:
          $ref: "#/components/schemas/PeriodSummary"
        costs:
          type: object
          additionalProperties: true
        volumes:
          type: object
          additionalProperties: true
        snapshots:
          type: object
          additionalProperties: true
    AskRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
        refresh:
          type: boolean
          default: false
    AskResponse:
      type: object
      properties:
        answer:
          type: string
          description: Markdown answer
        intent:
          type: string
        sources:
          type: array
          items:
            type: string
        ce_calls_used:
          type: integer
    ServiceDetail:
      type: object
      properties:
        account_id:
          type: string
        service:
          type: string
        start:
          type: string
          format: date
        end:
          type: string
          format: date
        total:
          type: number
        unit:
          type: string
        daily:
          type: array
          items:
            type: object
            additionalProperties: true
        by_usage_type:
          type: array
          items:
            type: object
            additionalProperties: true
        by_region:
          type: array
          items:
            type: object
            additionalProperties: true
        by_operation:
          type: array
          items:
            type: object
            additionalProperties: true
        by_name_tag:
          type: array
          items:
            type: object
            additionalProperties: true
    TagDelta:
      type: object
      properties:
        key:
          type: string
        current_usd:
          type: number
        prior_usd:
          type: number
        change_usd:
          type: number
        change_percent:
          type: number
        current_share_pct:
          type: number
    AccountTagDelta:
      type: object
      properties:
        account_id:
          type: string
        account_name:
          type: string
        total_current_usd:
          type: number
        total_prior_usd:
          type: number
        delta_usd:
          type: number
        delta_percent:
          type: number
        top_increases:
          type: array
          items:
            $ref: "#/components/schemas/TagDelta"
        top_decreases:
          type: array
          items:
            $ref: "#/components/schemas/TagDelta"
    ServiceTagDeltaResponse:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        service:
          type: string
        display_name:
          type: string
        tag_key:
          type: string
        current_period:
          $ref: "#/components/schemas/PeriodSummary"
        prior_period:
          $ref: "#/components/schemas/PeriodSummary"
        total_current_usd:
          type: number
        total_prior_usd:
          type: number
        delta_usd:
          type: number
        delta_percent:
          type: number
        top_increases:
          type: array
          items:
            $ref: "#/components/schemas/TagDelta"
        top_decreases:
          type: array
          items:
            $ref: "#/components/schemas/TagDelta"
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountTagDelta"
        top_account_note:
          type: string
        ce_calls_used:
          type: integer
    TagBucket:
      type: object
      properties:
        key:
          type: string
        current_usd:
          type: number
        current_share_pct:
          type: number
    ServiceTagTotalsResponse:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        service:
          type: string
        display_name:
          type: string
        tag_key:
          type: string
        current_period:
          $ref: "#/components/schemas/PeriodSummary"
        total_current_usd:
          type: number
        buckets:
          type: array
          items:
            $ref: "#/components/schemas/TagBucket"
        ce_calls_used:
          type: integer